+++ ggc/core/ParserVisitor.java
@@ -8,9 +8,9 @@ public interface ParserVisitor {
 	public void visitPartner(Partner partner) throws Exception;
 
 	/// Visits a bundle
-	public void visitBundle(String productId, String partnerId, int quantity, float unitPrice) throws Exception;
+	public void visitBundle(String productId, String partnerId, int quantity, double unitPrice) throws Exception;
 
 	/// Visits a bundle of derived products
-	public void visitDerivedBundle(String productId, String partnerId, int quantity, float unitPrice, float costFactor,
-			Map<String, Integer> recipeProducts) throws Exception;
+	public void visitDerivedBundle(String productId, String partnerId, int quantity, double unitPrice,
+			double costFactor, Map<String, Integer> recipeProducts) throws Exception;
 }
